Move gross single-macro-only "header" into wallet2
This header is stupid and gross and is only used from wallet2.  Remove
it from common/ to contain the damage within wallet2.

simplewallet.cpp includes it just for fun, but never actually uses it,
because why not?
